requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
distlib==0.3.1
Django==3.1.4
django-crispy-forms==1.10.0
django-phonenumber-field==5.0.0
filelock==3.0.12
flake8==3.8.4
gunicorn==20.0.4
idna==2.10
mccabe==0.6.1
phonenumbers==8.12.16
Pillow==8.1.0
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2019.3
requests==2.25.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.26.2
virtualenv==20.2.2
